+++ core/src/object/media/old_media_processor/shallow.rs
@@ -262,7 +262,7 @@ async fn get_files_for_labeling(
 				""
 			}
 		),
-		PrismaValue::Int(parent_iso_file_path.location_id() as i64),
+		PrismaValue::Int(parent_iso_file_path.location_id()),
 		PrismaValue::String(
 			parent_iso_file_path
 				.materialized_path_for_children()
